# spring-server/src/main/resources/api.yaml
swagger: '2.0'
info:
  description: API REST JavaLovers
  version: 1.0.0
  title: Specs of API JavaLovers
host: 'localhost:8080'
basePath: /
tags:
  - name: comments
    description: Operations about comments
schemes:
  - http
  - https
paths:
  /auth:
    post:
      tags:
        - auth
      summary: Authenticate an administrator.
      description: Authenticate an administrator
      operationId: authentication
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: authentication information
          required: true
          schema:
            $ref: '#/definitions/Credentials'
      responses:
        '200':
          description: OK - successful operation (Get JWT token header Authorization)
          headers:
            Authorization:
              description: the generated JWT is transmitted in this header
              type: string
        '400':
          description: BAD REQUEST - the received object is not of the expected type
        '401':
          description: UNAUTHORIZED - incorrect username or password
  /comments:
    post:
      tags:
        - comments
      summary: add a new comment
      description: add a new comment to the list of comments
      operationId: addComment
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Comment to add
          required: true
          schema:
            $ref: '#/definitions/Comment'
      responses:
        '201':
          description: successful operation
          headers:
            Location:
              type: string
        '400':
          description: the received object does not correspond to the expected resource
    get:
      tags:
        - comments
      summary: Get all comments
      description: Get the list of all visitor's comments
      operationId: getComments
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Comment'
  '/comments/{id}':
    get:
      tags:
        - comments
      summary: Get single comment
      description: Get a single comment (by ID)
      operationId: getCommentById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The comment that needs to be fetched.'
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Comment'
        '404':
          description: Comment not found
    delete:
      tags:
        - comments
      summary: Delete a comment
      description: Delete a comment (by ID)
      operationId: deleteCommentById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 'The comment that has to be deleted. (id)'
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
        '404':
          description: Comment not found
        '401':
          description: Unauthorized
definitions:
  Comment:
    type: object
    properties:
      name:
        type: string
        minLength: 3
        maxLength: 25
      text:
        type: string
        minLength: 1
        maxLength: 300
      date:
        type: string
        format: date-time
    required:
      - text
      - name
  Credentials:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
    required:
      - username
      - password